Explosive Bullet + Explosive Gun
Added an Explosive Gun item that has a random chance of dropping when an enemy is defeated. When picked up, it shoots explosive bullets that have a blast radius of 2f and does 3 damage. Item Ability Duration is 15s and returns to the Default Bullet when it is done.

The kind of item that drops is currently at random without a fixed probability yet. Also, items are able to stack (i.e. I could have a speed boost and explosive gun at the same time). Might be changed in the future where you could only use one item at a time.